fix(shared): Go 품질 게이트 차단 항목 정리
- hololive-shared의 nil, errcheck, scan, resource handling 경로를 Stage 3 lint/NilAway 기준에 맞게 수정

- alarm/domain/outbox/poller/scraper 계약 테스트와 helper를 보강

- workspace contract tests와 architecture import graph를 갱신

+++ hololive/hololive-shared/pkg/service/youtube/outbox/internal/delivery/dispatch/delivery_latency_period_rollups_test.go
@@ -271,7 +271,7 @@ func TestDeliveryTelemetryRepository_ListPostLatencyPeriodSummaries_UsesStoredPo
 	require.Equal(t, int64(2), summaries[1].ShortsPostCount)
 	require.Equal(t, int64(1), summaries[1].ShortsExceededPostCount)
 	require.NotNil(t, summaries[1].AverageLatencyMillis)
-	require.Equal(t, int64((withinLatencyMillis+exceededLatencyMillis)/2), *summaries[1].AverageLatencyMillis)
+	require.Equal(t, (withinLatencyMillis+exceededLatencyMillis)/2, *summaries[1].AverageLatencyMillis)
 	require.NotNil(t, summaries[1].P95LatencyMillis)
 	require.Equal(t, exceededLatencyMillis, *summaries[1].P95LatencyMillis)
 	require.NotNil(t, summaries[1].MaxLatencyMillis)
